# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "YEAH API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/data/pembayaran": {
            "get": {
                "tags": [
                    "Data"
                ],
                "summary": "ambil data pembayaran",
                "description": "ambil data pembayaran",
                "operationId": "pembayaran",
                "responses": {
                    "200": {
                        "description": "sukses ambil data pembayaran"
                    },
                    "400": {
                        "description": "gagal mendapatkan data pembayaran"
                    }
                }
            }
        },
...
    }
}